platform kuan 3

+++ user-dexon/contracts/Database.sol
@@ -25,20 +25,20 @@ contract CDatabase is CUserProfiles, CStatus, CTransaction {
         return txDatabaseSize;
     }
         
-    function setBuyTx(uint32 _txId, address _buyer) internal{
+    function setBuyTx(uint256 _txId, address _buyer) internal{
         txDatabase[_txId]._buyer = _buyer;
         _held(_buyer, txDatabase[_txId]._value);
         txDatabase[_txId]._status = Status.BUYING;
     }
-    function setPendTx(uint32 _txId, address _driver) internal{
+    function setPendTx(uint256 _txId, address _driver) internal{
         txDatabase[_txId]._driver = _driver;
         txDatabase[_txId]._status = Status.PENDING;
     }
-    function setDeliverTx(uint32 _txId) internal{
+    function setDeliverTx(uint256 _txId) internal{
         _held(txDatabase[_txId]._driver, txDatabase[_txId]._value * depositRatio);
         txDatabase[_txId]._status = Status.DELIVERING;
     }
-    function setSuccessTx(uint32 _txId) internal{
+    function setSuccessTx(uint256 _txId) internal{
         _unheld(txDatabase[_txId]._buyer, txDatabase[_txId]._value);
         _transfer(txDatabase[_txId]._buyer, txDatabase[_txId]._seller, txDatabase[_txId]._value);
         _unheld(txDatabase[_txId]._driver, txDatabase[_txId]._value * depositRatio);
